+++ private/plugins/mediagallery/language/polish_utf-8.php
@@ -891,7 +891,9 @@ $LANG_MG03 = array(
     'ip_rated' => 'Twój adres IP już oddał głos.',
     'uid_rated' => 'Twój ID już oddał głos.',
     'media_id' => 'Media ID',
-    'all_media' => 'All Media'
+    'all_media' => 'All Media',
+    'copy_to_clipboard' => 'Copy to Clipboard',
+    'media_id_ctc' => 'Media ID Copied to Clipboard'
 );
 
 $LANG_MG04 = array(